Manage the enabling of APIs directly in the initalisation script.
The main shift from moving the management of the APIs from terraform manifests to the initalisation script was due to unexpected dependency chains tying backing to the google_project_service resource.
One could indeed manage the APIs in terraform, but the explicit dependency DAG would be quite verbose and messy.
Not to mention that the terraform manifests become much more fragile and there is large amount of complexity creep that comes with it too.
I'm sure they is probably a better way to manage the direct enabling and management of google apis, but for the sake of this project and my skill level, this will do.
More research will be done to better automate and manage the underlying google apis.