Refactor the GitHub Actions OIDC/WIF into its own module.
The reason for this change is to drastically simplify and reduce the amount of business specific logic code in play.
Although it can be implemented on its own, OIDC/WIF external data can change (with allowed audiences, token urls etc).
No need to rewrite the wheel and it ensures everything is modular :).